pyproject.toml
[tool.poetry]
name = "fast-api-boiler"
version = "0.1.0"
description = ""
authors = ["Taha Aslam <tahaaslam045@gmail.com>"]
readme = "README.md"
packages = [{include = "fast_api_boiler"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.89.1"
sqlalchemy = "^1.4.46"
asyncpg = "^0.27.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
pydantic = {extras = ["email"], version = "^1.10.4"}
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
